Move chat additions to lab
