--- packages/cli/templates/blocks/components/ChatReactionBar/ChatReactionBarShowcase.doc.mjs
@@ -1,14 +1,0 @@
-// Copyright (c) Meta Platforms, Inc. and affiliates.
-
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  exampleFor: 'ChatReactionBar',
-  name: 'ChatReactionBar',
-  displayName: 'Chat Reaction Bar',
-  description: 'Emoji reaction pills under a chat message with toggleable own-reaction state and an add-reaction button that opens the emoji picker popover.',
-  isReady: true,
-  isShowcase: true,
-  aspectRatio: 16 / 9,
-  componentsUsed: ['Chat'],
-};
