--- packages/cli/templates/blocks/components/ChatTypingIndicator/ChatTypingIndicatorShowcase.doc.mjs
@@ -1,14 +1,0 @@
-// Copyright (c) Meta Platforms, Inc. and affiliates.
-
-/** @type {import('../../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'block',
-  exampleFor: 'ChatTypingIndicator',
-  name: 'ChatTypingIndicator',
-  displayName: 'Chat Typing Indicator',
-  description: 'Animated typing indicator below a conversation, shown alongside an unread divider marking where new messages begin.',
-  isReady: true,
-  isShowcase: true,
-  aspectRatio: 16 / 9,
-  componentsUsed: ['Chat'],
-};
