+++ packages/lab/src/Chat/ChatReactionBar.doc.mjs
@@ -1,6 +1,6 @@
 // Copyright (c) Meta Platforms, Inc. and affiliates.
 
-/** @type {import('../docs-types').ComponentDoc} */
+/** @type {import('../../core/src/docs-types').ComponentDoc} */
 
 export const docs = {
   name: 'ChatReactionBar',
