+++ packages/lab/src/Chat/ChatReactionBar.test.tsx
