+++ packages/lab/src/Chat/ChatTypingIndicator.test.tsx
