+++ packages/lab/src/Chat/ChatUnreadDivider.tsx
@@ -13,9 +13,9 @@
  * chat thread.
  *
  * SYNC: When modified, update these files to stay in sync:
- * - /packages/core/src/Chat/index.ts (exports)
- * - /packages/core/src/Chat/ChatUnreadDivider.doc.mjs
- * - /packages/core/src/Chat/Chat.doc.mjs (components list)
+ * - /packages/lab/src/Chat/index.ts (exports)
+ * - /packages/lab/src/Chat/ChatUnreadDivider.doc.mjs
+ * - /apps/storybook/stories/ChatAdditions.stories.tsx (examples)
  */
 
 import * as stylex from '@stylexjs/stylex';
@@ -25,10 +25,10 @@ import {
   typographyVars,
   typeScaleVars,
   fontWeightVars,
-} from '../theme/tokens.stylex';
-import {mergeProps} from '../utils';
-import {themeProps} from '../utils/themeProps';
-import type {BaseProps} from '../BaseProps';
+} from '@astryxdesign/core/theme/tokens.stylex';
+import type {BaseProps} from '@astryxdesign/core';
+import {mergeProps} from '@astryxdesign/core/utils';
+import {themeProps} from '@astryxdesign/core/utils';
 
 export interface ChatUnreadDividerProps extends BaseProps<HTMLDivElement> {
   /** Ref forwarded to the root element */
