+++ packages/themes/matcha/src/icons.tsx
@@ -30,7 +30,6 @@ import {
   Wrench,
   Square,
   Mic,
-  Smile,
 } from 'lucide-react';
 
 const iconProps = {
@@ -65,5 +64,4 @@ export const matchaIconRegistry: IconRegistry = {
   wrench: <Wrench {...iconProps} />,
   stop: <Square {...iconProps} />,
   microphone: <Mic {...iconProps} />,
-  faceSmile: <Smile {...iconProps} />,
 };
